Created DepartmentDeleteViewModel and made changes to Delete (GET and POST) action methods. In addition, I made necessary changes to Delete.cshtml

+++ ContosoUniversity/Controllers/DepartmentsController.cs
@@ -46,7 +46,7 @@ namespace ContosoUniversity.Controllers
         // GET: Departments/Create
         public IActionResult Create()
         {
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "FirstMidName");
+            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "LastName");
             return View();
         }
 
@@ -63,7 +63,7 @@ namespace ContosoUniversity.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "FirstMidName", department.InstructorID);
+            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "LastName", department.InstructorID);
             return View(department);
         }
 
@@ -169,6 +169,7 @@ namespace ContosoUniversity.Controllers
                     // Save the current RowVersion so next postback
                     // matches unless a new concurrency issue happens.
                     department.RowVersion = (byte[])dbValues.RowVersion;
+
                     // Clear the model error for the next postback.
                     ModelState.Remove("Department.RowVersion");
                 }
@@ -182,33 +183,59 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Departments/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? concurrencyError)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var department = await _context.Departments
+            DepartmentDeleteViewModel departmentDeleteVM = new DepartmentDeleteViewModel();
+
+            departmentDeleteVM.Department = await _context.Departments
                 .Include(d => d.Administrator)
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.DepartmentID == id);
-            if (department == null)
+
+            if (departmentDeleteVM.Department == null)
             {
                 return NotFound();
             }
 
-            return View(department);
+            if (concurrencyError.GetValueOrDefault())
+            {
+                departmentDeleteVM.ConcurrencyErrorMessage = "The record you attempted to delete "
+                  + "was modified by another user after you selected delete. "
+                  + "The delete operation was canceled and the current values in the "
+                  + "database have been displayed. If you still want to delete this "
+                  + "record, click the Delete button again.";
+            }
+
+            return View(departmentDeleteVM);
         }
 
         // POST: Departments/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int id, 
+            [Bind("DepartmentID,Name,Budget,StartDate,InstructorID,RowVersion")] Department department)
         {
-            var department = await _context.Departments.FindAsync(id);
-            _context.Departments.Remove(department);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                if (await _context.Departments.AnyAsync(m => m.DepartmentID == id))
+                {
+                    // Department.rowVersion value is from when the entity
+                    // was fetched. If it doesn't match the DB, a
+                    // DbUpdateConcurrencyException exception is thrown.
+                    _context.Departments.Remove(department);
+                    await _context.SaveChangesAsync();
+                }
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                return RedirectToAction(nameof(Delete), new { concurrencyError = true, id = id });
+            }
         }
 
         private bool DepartmentExists(int id)
